add types to remove errors
