+++ .eslintrc
@@ -1,56 +1,71 @@
 {
   "env": {
-      "browser": true,      
-      "es2020": true,
-      "node":true,
-      "es6":true
+    "browser": true,
+    "es2021": true,
+    "node": true,
+    "es6": true
   },
-  "extends": ["airbnb", "eslint:recommended", "plugin:import/recommended", "plugin:prettier/recommended", "plugin:jsx-a11y/recommended"],
-  "plugins": ["prettier" , "import", "jsx-a11y"],
+  "extends": [
+    "airbnb",
+    "eslint:recommended",
+    "plugin:import/recommended",
+    "plugin:prettier/recommended",
+    "plugin:jsx-a11y/recommended",
+    "plugin:@typescript-eslint/recommended",
+    "prettier"
+  ],
+  "plugins": ["import", "jsx-a11y", "@typescript-eslint"],
   "parserOptions": {
-      "ecmaVersion": "latest",
-      "sourceType": "module",
-      "ecmaFeatures": {
-        "globalReturn": false
-      },
-      "babelOptions": {
-        "configFile": "./babel.config.json"
-      },
-      "requireConfigFile": false
-      
+    "ecmaVersion": "latest",
+    "sourceType": "module",
+    "ecmaFeatures": {
+      "globalReturn": false
+    },
+    "babelOptions": {
+      "configFile": "./babel.config.json"
+    },
+    "requireConfigFile": false
   },
   "globals": {
     "Atomics": "readonly",
     "SharedArrayBuffer": "readonly"
-	},
-  "parser": "@babel/eslint-parser",
-  
-  "rules": {    
+  },
+  // "parser": "@babel/eslint-parser",
+  "parser": "@typescript-eslint/parser",
+
+  "rules": {
     "semi": "error",
-    "strict" : ["error","global"],
-    "func-style":["error", "expression"],
-    "no-new-func":"error",
-    "no-param-reassign": "error",    
+    "strict": ["error", "global"],
+    "func-style": ["error", "expression"],
+    "no-new-func": "error",
+    "no-param-reassign": "error",
     "prefer-arrow-callback": "warn",
     "arrow-parens": ["error", "always"],
     "arrow-body-style": ["error", "as-needed"],
     "new-cap": "error",
     "no-invalid-this": "error",
-    "prefer-const":"error",
-    "prefer-destructuring":["error",{"array":true, "object":true},{"enforceForRenamedProperties":true}],     
+    "prefer-const": "error",
+    "prefer-destructuring": [
+      "error",
+      { "array": true, "object": true },
+      { "enforceForRenamedProperties": true }
+    ],
     "prettier/prettier": "error",
     "no-unused-vars": "warn",
     "no-console": "warn",
     "func-names": "warn",
     "no-process-exit": "warn",
     "object-shorthand": "warn",
-    "class-methods-use-this": "error",    
-    "no-var": "error",    
-    "prefer-rest-params": "warn",    
-    "no-eval":"error",
-    "no-implied-eval":"error",
-    "eqeqeq":"error",
-    "no-with":"error",
-    "no-plusplus":"error"
+    "class-methods-use-this": "error",
+    "no-var": "error",
+    "prefer-rest-params": "warn",
+    "no-eval": "error",
+    "no-implied-eval": "error",
+    "eqeqeq": "error",
+    "no-with": "error",
+    "no-plusplus": "error",
+    "@typescript-eslint/no-unused-vars": "error",
+    // to enforce using type for object type definitions, can be type or interface
+    "@typescript-eslint/consistent-type-definitions": ["error", "type"]
   }
 }
