+++ src/fonts/amaranth-webfont/style.css
@@ -1,32 +1,30 @@
 /* #### Generated By: http://www.cufonfonts.com #### */
 
-    @font-face {
-    font-family: 'Amaranth Regular';
-    font-style: normal;
-    font-weight: normal;    
-    src: local('Amaranth Regular'), url('Amaranth-Regular.woff') format('woff');
-    }
-    
+@font-face {
+  font-family: "Amaranth Regular";
+  font-style: normal;
+  font-weight: normal;
+  src: local("Amaranth Regular"), url("Amaranth-Regular.woff") format("woff");
+}
 
-    @font-face {
-    font-family: 'Amaranth Italic';
-    font-style: normal;
-    font-weight: normal;
-    src: local('Amaranth Italic'), url('Amaranth-Italic.woff') format('woff');
-    }
-    
+@font-face {
+  font-family: "Amaranth Italic";
+  font-style: normal;
+  font-weight: normal;
+  src: local("Amaranth Italic"), url("Amaranth-Italic.woff") format("woff");
+}
 
-    @font-face {
-    font-family: 'Amaranth Bold';
-    font-style: normal;
-    font-weight: normal;
-    src: local('Amaranth Bold'), url('Amaranth-Bold.woff') format('woff');
-    }
-    
+@font-face {
+  font-family: "Amaranth Bold";
+  font-style: normal;
+  font-weight: normal;
+  src: local("Amaranth Bold"), url("Amaranth-Bold.woff") format("woff");
+}
 
-    @font-face {
-    font-family: 'Amaranth Bold Italic';
-    font-style: normal;
-    font-weight: normal;
-    src: local('Amaranth Bold Italic'), url('Amaranth-BoldItalic.woff') format('woff');
-    }
+@font-face {
+  font-family: "Amaranth Bold Italic";
+  font-style: normal;
+  font-weight: normal;
+  src: local("Amaranth Bold Italic"),
+    url("Amaranth-BoldItalic.woff") format("woff");
+}
