+++ src/vendor.ts
